Create BASH_it_all_SeqProcessing
This script should allow to fetch and process all necessary data for this project on the computer cluster at University of Bern.
For it to run, one must be logged into the cluster and run in this script from the terminal. If you are not part of this course, and thus do not have access to the same cluster, or if you would like to re-purpose this script for your own / a different analysis, make sure you change some of the cluster-, and user-specific parameters accordingly. For example, you might not have the same directory-layout & naming, and you might not have the same computing powers available.